requirements.txt
discord.py>=2.3.0
openai>=1.50.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiosqlite>=0.20.0
pyjwt>=2.9.0
python-multipart>=0.0.12
httpx>=0.27.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
